Add Context to ExecHarness executions
